Refactor `check_permissions` logic and move implementation to `security.py` for better modularity. Adjust imports and simplify `auth_client.py`.
