+++ neuro_admin_client/__init__.py
@@ -1,8 +1,9 @@
 from . import entities
 from .admin_client import AdminClient, GetUserResponse
-from .auth_client import AuthClient, check_permissions
+from .auth_client import AuthClient
 from .bearer_auth import BearerAuth
 from .entities import *  # noqa: F403
+from .security import check_permissions
 
 __all__ = [
     "AuthClient",
